Checking for Function
I had a major crash of Aptana/GitHub, where it lost nearly all of my CS
files, as well as deleting project files in Aptana so it didn't work.
I'm trying to see if all is well, and if this commit goes off without a
hitch, everything should be back in line.